backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
pvlib==0.10.3
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
scipy==1.11.4
requests==2.31.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
celery==5.3.4
geographiclib==2.0
shapely==2.0.1
folium==0.15.1
openpyxl==3.1.2
xlrd==2.0.1
pydantic>=2.0.0
pydantic-settings>=2.0.0
pydantic[email]>=2.0.0